+++ controller/index.js
@@ -1,0 +1,5 @@
+function index(req, res) {
+    res.render('index.ejs');
+}
+
+module.exports = index;
