+++ controller/login.js
@@ -1,0 +1,5 @@
+function login(req, res) {
+    res.render('login.ejs');
+}
+
+module.exports = login;
