+++ controller/pageNotFound.js
@@ -1,0 +1,5 @@
+function pageNotFound(req, res) {
+    res.status(404).render('not-found.ejs');
+}
+
+module.exports = pageNotFound;
