+++ controller/redirectLogin.js
@@ -1,0 +1,9 @@
+function redirectLogin(req, res, next) {
+    if (!req.session.user) {
+        res.redirect('/login');
+    } else {
+        next();
+    }
+}
+
+module.exports = redirectLogin;
