+++ controller/register.js
@@ -1,0 +1,5 @@
+function register(req, res) {
+    res.render('register.ejs');
+}
+
+module.exports = register;
